@@ -41,7 +41,7 @@
     "passport-google-oauth": "^2.0.0",
     "passwordless": "^1.1.3",
     "passwordless-tokenstore": "^0.0.10",
-    "socket.io": "^3.0.3",
+    "socket.io": "^4.2.0",
     "stripe": "^8.129.0",
     "typescript": "^4.1.3",
     "winston": "^3.3.3"
@@ -57,10 +57,10 @@
     "@types/node": "^12.12.2",
     "@types/node-fetch": "^1.6.9",
     "@types/passport": "^0.4.5",
-    "@types/socket.io": "^1.4.33",
+    "@types/socket.io": "^3.0.2",
     "@typescript-eslint/eslint-plugin": "^4.2.0",
     "@typescript-eslint/parser": "^4.2.0",
-    "eslint": "^6.8.0",
+    "eslint": "^7.25.0",
     "eslint-config-prettier": "^7.1.0",
     "eslint-plugin-prettier": "^3.3.1",
     "jest": "^26.4.2",
