@@ -14,41 +14,45 @@
     "start": "node production-server/server.js"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@material-ui/styles": "^4.11.4",
-    "@stripe/stripe-js": "^1.9.0",
-    "dotenv": "^8.2.0",
+    "@emotion/cache": "^11.6.0",
+    "@emotion/react": "^11.6.0",
+    "@emotion/server": "^11.4.0",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.0",
+    "@mui/material": "^5.2.0",
+    "@stripe/stripe-js": "^1.21.1",
     "express": "^4.17.1",
     "he": "^1.2.0",
     "isomorphic-unfetch": "^3.1.0",
     "keycode": "^2.2.0",
     "lru-cache": "^6.0.0",
     "marked": "^1.2.7",
-    "mobx": "^6.3.1",
-    "mobx-react": "6.3.1",
+    "mobx": "^6.3.7",
+    "mobx-react": "^7.2.1",
     "moment": "^2.29.1",
-    "next": "^9.1.2",
+    "next": "^12.0.3",
     "nprogress": "0.2.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-mentions": "3.0.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-mentions": "^4.3.0",
     "sitemap": "^6.3.5",
-    "socket.io-client": "^3.0.4",
-    "typescript": "^4.1.3"
+    "socket.io-client": "^4.2.0",
+    "typescript": "^4.4.4"
   },
   "devDependencies": {
+    "@babel/core": "^7.16.0",
+    "@babel/plugin-proposal-private-methods": "^7.16.0",
+    "@babel/plugin-proposal-private-property-in-object": "^7.16.0",
     "@types/express": "^4.11.1",
     "@types/he": "^0.5.29",
     "@types/marked": "^0.3.0",
     "@types/node": "^12.12.2",
     "@types/react": "^16.8.24",
     "@types/react-dom": "^16.0.5",
-    "@types/socket.io-client": "^1.4.33",
     "@typescript-eslint/eslint-plugin": "^4.2.0",
     "@typescript-eslint/parser": "^4.2.0",
-    "eslint": "^6.8.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.25.0",
     "eslint-config-prettier": "^7.1.0",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react": "^7.21.5",
