@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Autocomplete from '@material-ui/lab/Autocomplete';
-import TextField from '@material-ui/core/TextField';
+import Autocomplete from '@mui/material/Autocomplete';
+import TextField from '@mui/material/TextField';
 
 import { User } from '../../lib/store/user';
 
@@ -47,7 +47,7 @@ class MemberChooser extends React.Component<Props, State> {
         id="tags-standard"
         options={suggestions}
         getOptionLabel={(option) => option.label}
-        getOptionSelected={(option, value) => option.id === value.id}
+        isOptionEqualToValue={(option, value) => option.id === value.id}
         value={this.state.selectedItems}
         renderInput={(params) => (
           <TextField
