@@ -1,5 +1,5 @@
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
+import Menu from '@mui/material/Menu';
+import MenuItem from '@mui/material/MenuItem';
 import Link from 'next/link';
 import { NextRouter, withRouter } from 'next/router';
 import React from 'react';
