@@ -1,10 +1,10 @@
-import { createMuiTheme } from '@material-ui/core/styles';
+import { createTheme } from '@mui/material/styles';
 
-const themeDark = createMuiTheme({
+const themeDark = createTheme({
   palette: {
     primary: { main: '#238636' },
     secondary: { main: '#b62324' },
-    type: 'dark',
+    mode: 'dark',
     background: { default: '#0d1117' },
     text: {
       primary: '#c9d1d9',
@@ -18,11 +18,11 @@ const themeDark = createMuiTheme({
   },
 });
 
-const themeLight = createMuiTheme({
+const themeLight = createTheme({
   palette: {
     primary: { main: '#238636' },
     secondary: { main: '#b62324' },
-    type: 'light',
+    mode: 'light',
     background: { default: '#fff' },
     text: {
       primary: '#222',
