@@ -1,4 +1,4 @@
-import Avatar from '@material-ui/core/Avatar';
+import Avatar from '@mui/material/Avatar';
 import { observer } from 'mobx-react';
 import Error from 'next/error';
 import Head from 'next/head';
