@@ -105,9 +105,9 @@ class DiscussionActionMenu extends React.Component<Props, State> {
 
     const selectedDiscussion = currentTeam.discussions.find((d) => d._id === id);
 
-    const discussionUrl = `${dev ? process.env.URL_APP : process.env.PRODUCTION_URL_APP}/teams/${
-      currentTeam.slug
-    }/discussions/${selectedDiscussion.slug}`;
+    const discussionUrl = `${
+      dev ? process.env.NEXT_PUBLIC_URL_APP : process.env.NEXT_PUBLIC_PRODUCTION_URL_APP
+    }/teams/${currentTeam.slug}/discussions/${selectedDiscussion.slug}`;
 
     try {
       if (window.navigator) {
