@@ -25,16 +25,18 @@ export default async function sendRequestAndGetResponse(path, opts: any = {}) {
 
   const qs = (opts.qs && `?${makeQueryString(opts.qs)}`) || '';
 
-  // console.log(`before: ${process.env.URL_API}${path}${qs}`);
+  // console.log(`before: ${process.env.NEXT_PUBLIC_URL_API}${path}${qs}`);
 
   const response = await fetch(
     opts.externalServer
       ? `${path}${qs}`
-      : `${dev ? process.env.URL_API : process.env.PRODUCTION_URL_API}${path}${qs}`,
+      : `${
+          dev ? process.env.NEXT_PUBLIC_URL_API : process.env.NEXT_PUBLIC_PRODUCTION_URL_API
+        }${path}${qs}`,
     Object.assign({ method: 'POST', credentials: 'include' }, opts, { headers }),
   );
 
-  // console.log(`after: ${process.env.URL_API}${path}${qs}`);
+  // console.log(`after: ${process.env.NEXT_PUBLIC_URL_API}${path}${qs}`);
 
   // console.log(response.status);
   // console.log(response.statusText);
