@@ -37,7 +37,7 @@ app.prepare().then(() => {
       }
     }
 
-    res.redirect(`${dev ? process.env.URL_APP : process.env.PRODUCTION_URL_APP}/${redirectUrl}`);
+    res.redirect(`${dev ? process.env.NEXT_PUBLIC_URL_APP : process.env.NEXT_PUBLIC_PRODUCTION_URL_APP}/${redirectUrl}`);
   });
 
   // server.get('/api/v1/public/get-user', (_, res) => {
@@ -84,6 +84,6 @@ app.prepare().then(() => {
   // listen(handle: any, listeningListener?: () => void): http.Server;
   // "@types/express-serve-static-core", "version": "4.17.1"
   server.listen(port, () => {
-    console.log(`> Ready on ${dev ? process.env.URL_APP : process.env.PRODUCTION_URL_APP}`);
+    console.log(`> Ready on ${dev ? process.env.NEXT_PUBLIC_URL_APP : process.env.NEXT_PUBLIC_PRODUCTION_URL_APP}`);
   });
 });
